refactor: Response 객체 순환 참조 및 응집도를 고려하여 패키지 위치 변경
