+++ src/main/java/woowa/camp/jspcafe/repository/reply/ReplyRepository.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import woowa.camp.jspcafe.domain.Reply;
-import woowa.camp.jspcafe.service.dto.ReplyResponse;
+import woowa.camp.jspcafe.repository.dto.response.ReplyResponse;
 
 public interface ReplyRepository {
