+++ src/main/java/woowa/camp/jspcafe/service/UserService.java
@@ -9,7 +9,7 @@ import woowa.camp.jspcafe.domain.exception.UserException;
 import woowa.camp.jspcafe.repository.user.UserRepository;
 import woowa.camp.jspcafe.repository.dto.UserUpdateRequest;
 import woowa.camp.jspcafe.service.dto.RegistrationRequest;
-import woowa.camp.jspcafe.service.dto.UserResponse;
+import woowa.camp.jspcafe.repository.dto.response.UserResponse;
 import woowa.camp.jspcafe.infra.time.DateTimeProvider;
 
 public class UserService {
