+++ src/main/java/woowa/camp/jspcafe/web/servlet/article/ArticleEditServlet.java
@@ -18,7 +18,7 @@ import woowa.camp.jspcafe.domain.exception.ArticleException;
 import woowa.camp.jspcafe.domain.exception.UnAuthorizationException;
 import woowa.camp.jspcafe.repository.dto.ArticleUpdateRequest;
 import woowa.camp.jspcafe.service.ArticleService;
-import woowa.camp.jspcafe.service.dto.ArticleUpdateResponse;
+import woowa.camp.jspcafe.service.dto.response.ArticleUpdateResponse;
 import woowa.camp.jspcafe.web.utils.PathVariableExtractor;
 
 @WebServlet(name = "ArticleEditServlet", value = "/articles/edit/*")
