+++ src/main/java/woowa/camp/jspcafe/web/servlet/user/UserServlet.java
@@ -17,7 +17,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import woowa.camp.jspcafe.domain.User;
 import woowa.camp.jspcafe.service.UserService;
-import woowa.camp.jspcafe.service.dto.UserResponse;
+import woowa.camp.jspcafe.repository.dto.response.UserResponse;
 
 @WebServlet(name = "userServlet", value = {"/users", "/users/*"})
 public class UserServlet extends HttpServlet {
